+++ app/moves/controllers/index.js
@@ -1,4 +1,5 @@
-const get = require('./get')
+const detail = require('./detail')
+const list = require('./list')
 const Form = require('./form')
 const PersonalDetails = require('./personal-details')
 const Assessment = require('./assessment')
@@ -6,7 +7,8 @@ const MoveDetails = require('./move-details')
 const Save = require('./save')
 
 module.exports = {
-  get,
+  detail,
+  list,
   Form,
   PersonalDetails,
   Assessment,
