+++ app/moves/views/detail.njk
@@ -9,7 +9,7 @@
 {% block beforeContent %}
   {{ govukBackLink({
     text: "Back to dashboard",
-    href: "/"
+    href: "/moves"
   }) }}
 
   {{ super() }}
